+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to FileHarbor are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+---
+
+## [2.2.2] – 2026-04-22
+
+### Added
+- Added new runner for wiki
+- Added new admin API for list users
+- Added new param for user name on upload file
 
 ---